+++ src/app/locales/en/components/-storage/swift.js
@@ -1,9 +1,32 @@
 export default {
-  username: { name: 'Admin username' },
-  password: { name: 'Admin password' },
-  authUrl: { name: 'Auth URL' },
-  containerName: { name: 'Container name' },
-  tenantName: { name: 'Tenant name' },
+  username: {
+    name: 'Admin username',
+    tip: 'The Keystone authentication username.',
+  },
+  password: {
+    name: 'Admin password',
+    tip: 'The Keystone authentication password.',
+  },
+  projectName: {
+    name: 'Project name',
+    tip: 'The Keystone project name.',
+  },
+  userDomainName: {
+    name: 'User domain name',
+    tip: 'The Keystone user domain name.',
+  },
+  projectDomainName: {
+    name: 'Project domain name',
+    tip: 'The Keystone project domain name.',
+  },
+  authUrl: {
+    name: 'Auth URL',
+    tip: 'The URL to OpenStack Identity Service (Keystone) V3.',
+  },
+  containerName: {
+    name: 'Container name',
+    tip: 'The name of the Swift storage container.',
+  },
   blockSize: { name: 'Block size' },
   timeout: { name: 'Timeout [ms]' },
 };